Added security to product api

+++ djv-product/src/main/java/lt/dejavu/product/api/ProductApi.java
@@ -1,12 +1,17 @@
 package lt.dejavu.product.api;
 
+import lt.dejavu.auth.exception.ApiSecurityException;
+import lt.dejavu.auth.model.Endpoint;
+import lt.dejavu.auth.service.SecurityService;
 import lt.dejavu.product.model.rest.request.ProductRequest;
 import lt.dejavu.product.dto.ProductDto;
 import lt.dejavu.product.service.ProductService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpHeaders;
 import org.springframework.http.MediaType;
 import org.springframework.web.bind.annotation.*;
 
+import javax.servlet.http.HttpServletRequest;
 import java.util.List;
 
 @RestController
@@ -15,6 +20,9 @@ public class ProductApi {
 
     @Autowired
     private ProductService productService;
+
+    @Autowired
+    private SecurityService securityService;
 
     @GetMapping(
             path="/",
@@ -45,7 +53,10 @@ public class ProductApi {
             produces = MediaType.APPLICATION_JSON_UTF8_VALUE
     )
 
-    public Long createProduct(@RequestBody ProductRequest productRequest){
+    public Long createProduct(HttpServletRequest request,
+                              @RequestHeader(HttpHeaders.AUTHORIZATION) String authHeader,
+                              @RequestBody ProductRequest productRequest) throws ApiSecurityException {
+        authorize(authHeader, request);
         return productService.createProduct(productRequest);
     }
 
@@ -53,7 +64,11 @@ public class ProductApi {
             path = "/{productId}",
             produces = MediaType.APPLICATION_JSON_UTF8_VALUE
     )
-    public void updateProduct(@PathVariable("productId") long productId, @RequestBody ProductRequest productRequest){
+    public void updateProduct(HttpServletRequest request,
+                              @RequestHeader(HttpHeaders.AUTHORIZATION) String authHeader,
+                              @PathVariable("productId") long productId,
+                              @RequestBody ProductRequest productRequest) throws ApiSecurityException {
+        authorize(authHeader, request);
         productService.updateProduct(productId, productRequest);
     }
 
@@ -61,8 +76,21 @@ public class ProductApi {
             path = "/{productId}",
             produces = MediaType.APPLICATION_JSON_UTF8_VALUE
     )
-    public void deleteProduct(@PathVariable("productId") long productId){
+    public void deleteProduct(HttpServletRequest request,
+                              @RequestHeader(HttpHeaders.AUTHORIZATION) String authHeader,
+                              @PathVariable("productId") long productId) throws ApiSecurityException {
+        authorize(authHeader, request);
         productService.deleteProduct(productId);
     }
 
+    private Endpoint buildEndpoint(HttpServletRequest request) {
+        Endpoint endpoint = new Endpoint();
+        endpoint.setMethod(RequestMethod.valueOf(request.getMethod()));
+        endpoint.setPath(request.getRequestURI());
+        return endpoint;
+    }
+
+    private long authorize(String authHeader, HttpServletRequest request) throws ApiSecurityException {
+        return securityService.authorize(authHeader, buildEndpoint(request));
+    }
 }
